Start, get election and database automatic migrations

+++ UrnaEletronicaBACK/UrnaEletronica.Application/DTOs/StartElectionDto.cs
@@ -5,7 +5,7 @@ namespace UrnaEletronica.Application.DTOs
     public class StartElectionDto
     {
         [Range(1, short.MaxValue, ErrorMessage = "Invalid election year")]
-        public ushort ElectionYear { get; set; }
+        public int ElectionYear { get; set; }
         public ICollection<StartElectionCycleDto>? ElectionCycles { get; set; }
     }
 }
